refactor(service): move homepage clean-up to homepage domain service
